Namespace the definitions
