--- lib/yuriita/single_definition.rb
@@ -1,28 +1,0 @@
-module Yuriita
-  class SingleDefinition
-    attr_reader :options
-
-    def initialize(options:)
-      @options = options
-    end
-
-    def apply(query:)
-      selector = SingleSelect.new(options: options, query: query)
-
-      filters = [selector.filter].compact
-      FilterClause.new(filters: filters, combination: combination)
-    end
-
-    def view_options(query:, param_key:)
-      SingleCollection.new(
-        definition: self,
-        query: query,
-        formatter: Yuriita::QueryFormatter.new(param_key: param_key),
-      ).view_options
-    end
-
-    def combination
-      AndCombination
-    end
-  end
-end
